GPS: using 16-bit integers for altitude and speed

+++ openrtx/include/core/gps.h
@@ -50,8 +50,8 @@ typedef struct
     uint32_t   active_sats;          // Bitmap representing which sats are part of the fix
     int32_t    latitude;             // Latitude coordinates
     int32_t    longitude;            // Longitude coordinates
-    float      altitude;             // Antenna altitude above mean sea level (geoid) in m
-    float      speed;                // Ground speed in km/h
+    int16_t    altitude;             // Antenna altitude above mean sea level (geoid) in m
+    uint16_t   speed;                // Ground speed in km/h
     float      tmg_mag;              // Course over ground, degrees, magnetic
     float      tmg_true;             // Course over ground, degrees, true
 }
